Improved UDF parameter verification and error messages
