+++ exasol/mlflow_plugin/rest_api/__init__.py
@@ -1,7 +1,9 @@
+from exasol.mlflow_plugin.rest_api.endpoints.endpoint import Endpoint
 from exasol.mlflow_plugin.rest_api.endpoints.experiments import EXPERIMENTS_SEARCH
 from exasol.mlflow_plugin.rest_api.udf.body import UdfBody
 
 __all__ = [
+    "Endpoint",
     "EXPERIMENTS_SEARCH",
     "UdfBody",
 ]
